+++ lib/warnings/plugin.rb
@@ -1,35 +1,38 @@
-require_relative 'report/reporter'
+# frozen_string_literal: true
+
+require_relative 'reporter/reporter_factory'
 
 module Danger
   # Create uniform issue reports for different parser types.
-  # @example Create a basic bandit report.
-  #       warnings.report(
+  # @example Create a basic bandit reporter.
+  #       warnings.reporter(
   #         name: 'Bandit Report',
-  #         parser: :bandit,
-  #         file: report/bandit.json
+  #         id: :bandit,
+  #         file: reporter/bandit.json
   #       )
   #
-  # @example Create a bandit report and comment inline for all files.
-  #       warnings.report(
+  # @example Create a bandit reporter and comment inline for all files.
+  #       warnings.reporter(
   #         name: 'Bandit Report',
-  #         parser: :bandit,
-  #         file: report/bandit.json,
+  #         id: :bandit,
+  #         file: reporter/bandit.json,
   #         inline: true,
   #         filter: false
   #       )
   #
   # @see Kyaak/danger-warnings
-  # @tags warnings, danger, parser, issues, report
+  # @tags warnings, danger, parser, issues, reporter
   class DangerWarnings < Plugin
-    # Whether to comment as markdown report or do an inline comment on the file.
+    ERROR_ID_NOT_SET = 'danger.warnings requires an :id'
+    # Whether to comment as markdown reporter or do an inline comment on the file.
     #
     # This will be set as default for all reporters used in this danger run.
     # It can still be overridden by setting the value when using #report.
     #
     # @return [Bool] Use inline comments.
     attr_accessor :inline
-    # Whether to filter and report only for changes files.
-    # If this is set to false, all issues are of a report are included in the comment.
+    # Whether to filter and reporter only for changes files.
+    # If this is set to false, all issues are of a reporter are included in the comment.
     #
     # This will be set as default for all reporters used in this danger run.
     # It can still be overridden by setting the value when using #report.
@@ -48,10 +51,10 @@ module Danger
       super(dangerfile)
     end
 
-    # Create a report for given arguments.
+    # Create a reporter for given arguments.
     #         name: 'Bandit Report',
     #         parser: :bandit,
-    #         file: report/bandit.json,
+    #         file: reporter/bandit.json,
     #         inline: true,
     #         filter: false
     # @param args List of arguments to be used.
@@ -67,8 +70,11 @@ module Danger
 
     # rubocop:disable Metrics/AbcSize
     def create_reporter(options)
-      reporter = Warnings::Reporter.new(self)
-      reporter.parser = options[:parser]
+      id = options[:id]
+      raise ERROR_ID_NOT_SET if id.nil?
+
+      reporter = Warnings::ReporterFactory.create(id, self)
+      reporter.format = options[:format]
       reporter.name = options[:name]
       reporter.file = options[:file]
       reporter.baseline = options[:baseline]