+++ spec/spec_helper/assets.rb
@@ -1,15 +1,17 @@
+# frozen_string_literal: true
+
 module Warnings
   module Assets
     ASSETS_DIR = Pathname.new(File.expand_path('../assets', __dir__))
-    EMPTY_FILE = "#{ASSETS_DIR}/empty.txt".freeze
+    EMPTY_FILE = "#{ASSETS_DIR}/empty.txt"
 
-    BANDIT_JSON = "#{ASSETS_DIR}/bandit.json".freeze
-    BANDIT_EMPTY = "#{ASSETS_DIR}/bandit_empty.json".freeze
-    BANDIT_MISSING_RESULTS = "#{ASSETS_DIR}/bandit_missing_results.json".freeze
-    PYLINT_TXT = "#{ASSETS_DIR}/pylint.txt".freeze
-    RUBOCOP_JSON = "#{ASSETS_DIR}/rubocop.json".freeze
-    RUBOCOP_MULTI_JSON = "#{ASSETS_DIR}/rubocop_multi_offenses.json".freeze
-    RUBOCOP_SIMPLE = "#{ASSETS_DIR}/rubocop.txt".freeze
+    BANDIT_JSON = "#{ASSETS_DIR}/bandit_json.json"
+    BANDIT_EMPTY = "#{ASSETS_DIR}/bandit_json_empty.json"
+    BANDIT_MISSING_RESULTS = "#{ASSETS_DIR}/bandit_json_missing_results.json"
+    PYLINT_PARSEABLE = "#{ASSETS_DIR}/pylint_parseable.txt"
+    RUBOCOP_JSON = "#{ASSETS_DIR}/rubocop_json.json"
+    RUBOCOP_MULTI_JSON = "#{ASSETS_DIR}/rubocop_json_multi_offenses.json"
+    RUBOCOP_SIMPLE = "#{ASSETS_DIR}/rubocop_simple.txt"
 
     BANDIT_FIRST_ISSUE = {
       code: "2852         except ImportError:\n2853             import pickle\n2854         with open(filename, 'wb') as outf:\n",