+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Refactor reporter and parser handling
+- Add `format` attribute
+- Rename `parser` to `id`
+
+### Removed
+- Ruby 2.2 support
 
 ## [0.1.0] - 2019-01-23
 ### Added